Create ruff-project crate
