+++ crates/ruff_cli/src/args.rs
@@ -8,12 +8,12 @@ use rustc_hash::FxHashMap;
 use ruff::line_width::LineLength;
 use ruff::logging::LogLevel;
 use ruff::registry::Rule;
-use ruff::resolver::ConfigProcessor;
 use ruff::settings::configuration::RuleSelection;
 use ruff::settings::types::{
     FilePattern, PatternPrefixPair, PerFileIgnore, PythonVersion, SerializationFormat,
 };
 use ruff::RuleSelector;
+use ruff_workspace::resolver::ConfigProcessor;
 
 #[derive(Debug, Parser)]
 #[command(
