+++ crates/ruff_workspace/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "ruff_workspace"
+version = "0.0.0"
+publish = false
+authors = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+
+[lib]
+
+[dependencies]
+ruff = { path = "../ruff" }
+
+anyhow = { workspace = true }
+ignore = { workspace = true }
+itertools = { workspace = true }
+log = { workspace = true }
+globset = { workspace = true }
+rustc-hash = { workspace = true }
+
+
+[dev-dependencies]
+tempfile = "3.6.0"
+path-absolutize = { workspace = true }
