+++ crates/ruff_workspace/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod resolver;
+
+#[cfg(test)]
+mod tests {
+    use std::path::Path;
+
+    pub(crate) fn test_resource_path(path: impl AsRef<Path>) -> std::path::PathBuf {
+        Path::new("../ruff/resources/test/").join(path)
+    }
+}
